+++ src/OperasiMatrix.java
@@ -1,7 +1,7 @@
 public class OperasiMatrix {
     // Membuat matrix A
-    public static float[][] makeMatrixA(float[][] matrix, int row, int col) {
-        float[][] MatrikA = new float[row][col];
+    public static double[][] makeMatrixA(double[][] matrix, int row, int col) {
+        double[][] MatrikA = new double[row][col];
         int idxRowNew, idxColNew;
     
         for (idxRowNew = 0; idxRowNew < row; idxRowNew++) {
@@ -12,8 +12,8 @@ public class OperasiMatrix {
         return MatrikA;
     }
     // Membuat matrix B
-    public static float[][] makeMatrixB(float[][] matrix, int row, int col) {
-        float[][] MatrikB = new float[row][1];
+    public static double[][] makeMatrixB(double[][] matrix, int row, int col) {
+        double[][] MatrikB = new double[row][1];
         int idxRowNew, idxColNew;
     
         for (idxRowNew = 0; idxRowNew < row; idxRowNew++) {
@@ -24,22 +24,22 @@ public class OperasiMatrix {
         return MatrikB;
     }
     // Mendapat baris matrix
-    public static int getRow(float[][] matrix) {
+    public static int getRow(double[][] matrix) {
             
         return matrix.length;
     }
     // Mendapatkan kolom matrix
-    public static int getCol(float[][] matrix) {
+    public static int getCol(double[][] matrix) {
             
         return matrix[0].length;
     }
     //Mengecek persegi atau Tidak
-    public static boolean isSquare(float[][] matrix) {
+    public static boolean isSquare(double[][] matrix) {
 
         return (getRow(matrix) == getCol(matrix));
     }
     //Mengeck matrixB nol semua atau tidak
-    public static boolean isMatrixBNol(float[][] matrix) {
+    public static boolean isMatrixBNol(double[][] matrix) {
         boolean isNol = true;
         int idx = 0;
 
@@ -53,14 +53,14 @@ public class OperasiMatrix {
         return isNol;
     }
     // Mengkalikan matirx
-    public static float[][] mutiplyMatrix (float[][] matrixA, float[][] matrixB, int row, int col) {
+    public static double[][] mutiplyMatrix (double[][] matrixA, double[][] matrixB, int row, int col) {
         int rowHasil = row, colHasil = col, idxRow, idxCol, idxPlus;
-        float[][] matrixHasil = new float[rowHasil][colHasil];
+        double[][] matrixHasil = new double[rowHasil][colHasil];
 
         for (idxRow = 0; idxRow < rowHasil; idxRow++) {
             int idxTemp = 0;
             for (idxCol = 0; idxCol < colHasil; idxCol++) {
-                float temp = 0;
+                double temp = 0;
                 int idxTempPlus = idxRow;
                 for (idxPlus = 0; idxPlus < matrixA[0].length; idxPlus++) {
                     temp += matrixA[idxTempPlus][idxPlus]*matrixB[idxPlus][idxTemp];
@@ -71,7 +71,7 @@ public class OperasiMatrix {
         }
         return matrixHasil;
     }
-    public static void changeCol(float[][] matrixA, float[][] matrixB, int col) {
+    public static void changeCol(double[][] matrixA, double[][] matrixB, int col) {
         int row = matrixA.length, idxRow;
 
         for (idxRow = 0; idxRow < row; idxRow++) {
@@ -79,3 +79,5 @@ public class OperasiMatrix {
         }
     }
 }
+
+
